requirements.txt
streamlit==1.32.0
requests==2.31.0
opencv-python-headless==4.9.0.80
pillow==10.2.0
numpy==1.26.4